separate len-vector class
