+++ Makefile
@@ -15,7 +15,7 @@ BINDIR := $(BUILD)/.o
 LDFLAGS := -pthread
 LIBS = -lz
 
-OBJS := main input reads
+OBJS := main input reads len-vector
 BINS := $(addprefix $(BINDIR)/, $(OBJS))
 
 #gfalibs
